backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2 